+++ erwin-component-sample-parent/sample-redis/src/main/java/com/fengwenyi/erwin/component/sample/redis/SampleRedisApplication.java
@@ -1,0 +1,17 @@
+package com.fengwenyi.erwin.component.sample.redis;
+
+import org.springframework.boot.SpringApplication;
+import org.springframework.boot.autoconfigure.SpringBootApplication;
+
+/**
+ * @author <a href="https://fengwenyi.com">Erwin Feng</a>
+ * @since 2023-12-29
+ */
+@SpringBootApplication
+public class SampleRedisApplication {
+
+    public static void main(String[] args) {
+        SpringApplication.run(SampleRedisApplication.class, args);
+    }
+
+}