+++ CHANGE_LOG.md
@@ -1,3 +1,8 @@
 # 更新记录
 
 ## 0.0.1
+
+- 【component-common-common】
+- 【component-common-jpa】
+- 【component-common-redis】
+- 【component-task】